requirements.txt
accelerate==0.27.0
numpy==1.26.1
openai==0.28.1
pandas==2.1.2
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.5.1
sentencepiece==0.1.99
tinycss2==1.2.1
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2

